# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Go 语言开发的会员预约系统 API 文档",
        "title": "会员预约系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "用户登录接口，支持用户名或手机号登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/member-pre_internal_domain_auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/member-pre_internal_domain_auth.LoginResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户登出接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登出",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户的信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "获取当前用户",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/member-pre_internal_domain_auth.GetCurrentUserResponse"
                        }
                    }
                }
            }
        },
        "/stores/{store_id}/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据门店ID获取该门店的用户列表（店长和美甲师）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "根据门店获取用户列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "门店ID",
                        "name": "store_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "store_manager",
                            "technician"
                        ],
                        "type": "string",
                        "description": "角色筛选",
                        "name": "role",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/member-pre_internal_domain_user.User"
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据用户ID获取用户详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "获取用户详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/member-pre_internal_domain_user.User"
                        }
                    }
                }
            }
        },
        "/users/{id}/work-status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新美甲师的工作状态（在岗/休息/离岗）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "更新美甲师工作状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "工作状态请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.UpdateWorkStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_interfaces_http_handler.UpdateWorkStatusRequest": {
            "type": "object",
            "required": [
                "work_status"
            ],
            "properties": {
                "work_status": {
                    "type": "string",
                    "enum": [
                        "working",
                        "rest",
                        "offline"
                    ],
                    "example": "working"
                }
            }
        },
        "member-pre_internal_domain_auth.GetCurrentUserResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/member-pre_internal_domain_user.User"
                }
            }
        },
        "member-pre_internal_domain_auth.LoginRequest": {
            "type": "object",
            "required": [
                "username"
            ],
            "properties": {
                "password": {
                    "description": "密码（员工必填，顾客不需要）",
                    "type": "string"
                },
                "username": {
                    "description": "用户名（员工）或手机号（顾客）",
                    "type": "string"
                }
            }
        },
        "member-pre_internal_domain_auth.LoginResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/member-pre_internal_domain_user.User"
                }
            }
        },
        "member-pre_internal_domain_user.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "phone": {
                    "description": "手机号（顾客和员工都可能有）",
                    "type": "string"
                },
                "role": {
                    "description": "角色: admin, store_manager, technician, customer",
                    "type": "string"
                },
                "status": {
                    "description": "状态: active, inactive",
                    "type": "string"
                },
                "store_id": {
                    "description": "门店ID（店长和美甲师必须关联，顾客和总后台为nil）",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "description": "用户名（员工使用）",
                    "type": "string"
                },
                "work_status": {
                    "description": "工作状态（仅美甲师有效）: working, rest, offline",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT认证，格式: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}